idms-analyzer/requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
python-docx==1.1.0
requests==2.31.0
backoff==2.2.1
pydantic==2.6.1
